+++ PoliNetworkBot_CSharp/Code/Utils/UtilsMedia/UtilsVideo.cs
@@ -66,7 +66,7 @@ namespace PoliNetworkBot_CSharp.Code.Utils.UtilsMedia
 
             try
             {
-                return Convert.ToInt32(r2);
+                return Convert.ToInt64(r2);
             }
             catch
             {
@@ -89,7 +89,7 @@ namespace PoliNetworkBot_CSharp.Code.Utils.UtilsMedia
 
             try
             {
-                return Convert.ToInt32(r2);
+                return Convert.ToInt64(r2);
             }
             catch
             {
@@ -107,10 +107,10 @@ namespace PoliNetworkBot_CSharp.Code.Utils.UtilsMedia
 
             var dr = dt.Rows[0];
 
-            return new ObjectVideo(Convert.ToInt32(dr["id_video"]), dr["file_id"].ToString(),
-                Convert.ToInt32(dr["file_size"]), Convert.ToInt32(dr["height"]),
-                Convert.ToInt32(dr["width"]), dr["unique_id"].ToString(),
-                messageIdFrom, chatIdFromIdPerson, chatType, Convert.ToInt32(dr["duration"]));
+            return new ObjectVideo((int)Convert.ToInt64(dr["id_video"]), dr["file_id"].ToString(),
+               (int)Convert.ToInt64(dr["file_size"]), (int)Convert.ToInt64(dr["height"]),
+              (int)Convert.ToInt64(dr["width"]), dr["unique_id"].ToString(),
+                messageIdFrom, chatIdFromIdPerson, chatType, (int)Convert.ToInt64(dr["duration"]));
         }
     }
 }
